requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-groq
requests
beautifulsoup4
python-dotenv
passlib
psycopg2-binary
PyJWT



